# core/vm: optimize jumpdest analysis