@holiman 2021-08-31
Here's what I get -- didn't run all of them, but `--count 10` on the ones I tested
```
name                        old time/op    new time/op    delta
JumpdestOpAnalysis/PUSH1-6    1.02ms ± 1%    1.03ms ± 1%     ~     (p=0.059 n=9+8)
JumpdestOpAnalysis/PUSH2-6     938µs ± 2%    1066µs ±10%  +13.61%  (p=0.000 n=10+10)
JumpdestOpAnalysis/STOP-6      814µs ± 7%     747µs ± 3%   -8.26%  (p=0.000 n=10+10)

name                        old speed      new speed      delta
JumpdestOpAnalysis/PUSH1-6  1.20GB/s ± 1%  1.19GB/s ± 1%     ~     (p=0.059 n=9+8)
JumpdestOpAnalysis/PUSH2-6  1.31GB/s ± 2%  1.16GB/s ± 9%  -11.74%  (p=0.000 n=10+10)
JumpdestOpAnalysis/STOP-6   1.51GB/s ± 7%  1.65GB/s ± 3%   +8.77%  (p=0.000 n=10+10)
```