@MariusVanDerWijden 2021-09-01
I can see the same regression on my machine
```
JumpdestOpAnalysis/PUSH1-24      751µs ±12%     823µs ± 4%   +9.65%  (p=0.007 n=10+10)
JumpdestOpAnalysis/PUSH2-24      722µs ± 6%     823µs ± 3%  +14.11%  (p=0.000 n=10+10)
JumpdestOpAnalysis/PUSH10-24     310µs ± 3%     298µs ± 4%   -3.74%  (p=0.000 n=10+10)
JumpdestOpAnalysis/PUSH11-24     293µs ± 2%     295µs ± 3%     ~     (p=0.393 n=10+10)
JumpdestOpAnalysis/PUSH12-24     255µs ± 2%     249µs ± 2%   -2.48%  (p=0.000 n=9+10)
JumpdestOpAnalysis/PUSH13-24     261µs ± 3%     253µs ± 1%   -3.20%  (p=0.000 n=10+8)
JumpdestOpAnalysis/PUSH14-24     252µs ± 4%     219µs ± 4%  -12.92%  (p=0.000 n=10+10)
JumpdestOpAnalysis/PUSH15-24     257µs ± 2%     242µs ± 4%   -6.07%  (p=0.000 n=10+10)
JumpdestOpAnalysis/PUSH16-24     213µs ± 2%     207µs ± 4%   -2.55%  (p=0.004 n=10+10)
JumpdestOpAnalysis/PUSH17-24     171µs ± 3%     154µs ± 3%  -10.26%  (p=0.000 n=10+10)
JumpdestOpAnalysis/PUSH18-24     196µs ± 3%     188µs ± 3%   -3.89%  (p=0.000 n=10+10)
JumpdestOpAnalysis/PUSH19-24     195µs ± 5%     181µs ± 2%   -7.42%  (p=0.000 n=10+10)
JumpdestOpAnalysis/PUSH20-24     191µs ± 4%     171µs ± 4%  -10.42%  (p=0.000 n=10+10)
JumpdestOpAnalysis/PUSH21-24     187µs ± 4%     176µs ± 3%   -6.07%  (p=0.000 n=10+10)
JumpdestOpAnalysis/PUSH22-24     182µs ± 4%     158µs ± 2%  -13.41%  (p=0.000 n=10+10)
JumpdestOpAnalysis/PUSH23-24     184µs ± 2%     177µs ± 1%   -3.93%  (p=0.000 n=10+9)
JumpdestOpAnalysis/PUSH24-24     177µs ± 5%     166µs ± 2%   -6.46%  (p=0.000 n=10+10)
JumpdestOpAnalysis/PUSH25-24     150µs ± 3%     142µs ± 4%   -5.67%  (p=0.000 n=10+9)
JumpdestOpAnalysis/PUSH26-24     174µs ± 3%     157µs ± 3%   -9.85%  (p=0.000 n=10+10)
JumpdestOpAnalysis/PUSH27-24     165µs ± 4%     159µs ± 1%   -3.67%  (p=0.001 n=10+9)
JumpdestOpAnalysis/PUSH28-24     161µs ± 3%     145µs ± 1%  -10.00%  (p=0.000 n=10+9)
JumpdestOpAnalysis/PUSH29-24     163µs ± 4%     141µs ± 3%  -13.60%  (p=0.000 n=10+10)
JumpdestOpAnalysis/STOP-24       532µs ± 4%     549µs ± 2%   +3.23%  (p=0.003 n=10+10)

name                          old speed      new speed      delta
JumpdestOpAnalysis/PUSH1-24   1.65GB/s ±12%  1.49GB/s ± 4%   -9.34%  (p=0.007 n=10+10)
JumpdestOpAnalysis/PUSH2-24   1.70GB/s ± 6%  1.49GB/s ± 3%  -12.40%  (p=0.000 n=10+10)
JumpdestOpAnalysis/PUSH10-24  3.97GB/s ± 3%  4.13GB/s ± 3%   +3.89%  (p=0.000 n=10+10)
JumpdestOpAnalysis/PUSH11-24  4.20GB/s ± 2%  4.16GB/s ± 3%     ~     (p=0.393 n=10+10)
JumpdestOpAnalysis/PUSH12-24  4.80GB/s ± 3%  4.95GB/s ± 2%   +2.93%  (p=0.000 n=10+10)
JumpdestOpAnalysis/PUSH13-24  4.70GB/s ± 3%  4.86GB/s ± 1%   +3.28%  (p=0.000 n=10+8)
JumpdestOpAnalysis/PUSH14-24  4.88GB/s ± 4%  5.60GB/s ± 4%  +14.83%  (p=0.000 n=10+10)
JumpdestOpAnalysis/PUSH15-24  4.77GB/s ± 2%  5.08GB/s ± 4%   +6.50%  (p=0.000 n=10+10)
JumpdestOpAnalysis/PUSH16-24  5.78GB/s ± 2%  5.93GB/s ± 3%   +2.63%  (p=0.004 n=10+10)
JumpdestOpAnalysis/PUSH17-24  7.18GB/s ± 3%  8.00GB/s ± 3%  +11.44%  (p=0.000 n=10+10)
JumpdestOpAnalysis/PUSH18-24  6.28GB/s ± 3%  6.53GB/s ± 3%   +4.05%  (p=0.000 n=10+10)
JumpdestOpAnalysis/PUSH19-24  6.26GB/s ± 2%  6.79GB/s ± 2%   +8.57%  (p=0.000 n=9+10)
JumpdestOpAnalysis/PUSH20-24  6.43GB/s ± 4%  7.18GB/s ± 4%  +11.61%  (p=0.000 n=10+10)
JumpdestOpAnalysis/PUSH21-24  6.58GB/s ± 4%  7.00GB/s ± 3%   +6.42%  (p=0.000 n=10+10)
JumpdestOpAnalysis/PUSH22-24  6.75GB/s ± 4%  7.80GB/s ± 2%  +15.47%  (p=0.000 n=10+10)
JumpdestOpAnalysis/PUSH23-24  6.69GB/s ± 2%  6.96GB/s ± 1%   +4.08%  (p=0.000 n=10+9)
JumpdestOpAnalysis/PUSH24-24  6.93GB/s ± 5%  7.41GB/s ± 2%   +6.88%  (p=0.000 n=10+10)
JumpdestOpAnalysis/PUSH25-24  8.19GB/s ± 3%  8.68GB/s ± 3%   +6.00%  (p=0.000 n=10+9)
JumpdestOpAnalysis/PUSH26-24  7.06GB/s ± 3%  7.83GB/s ± 3%  +10.93%  (p=0.000 n=10+10)
JumpdestOpAnalysis/PUSH27-24  7.43GB/s ± 4%  7.71GB/s ± 1%   +3.76%  (p=0.001 n=10+9)
JumpdestOpAnalysis/PUSH28-24  7.62GB/s ± 3%  8.47GB/s ± 1%  +11.10%  (p=0.000 n=10+9)
JumpdestOpAnalysis/PUSH29-24  7.55GB/s ± 4%  8.74GB/s ± 3%  +15.69%  (p=0.000 n=10+10)
JumpdestOpAnalysis/STOP-24    2.31GB/s ± 4%  2.24GB/s ± 2%   -3.17%  (p=0.003 n=10+10)
```