@holiman 2022-06-03
Yes, I agree

```
[user@work vm]$ benchstat before.txt after.txt 
name                        old time/op   new time/op   delta
JumpdestOpAnalysis/PUSH1-6   1.58ms ±19%   1.46ms ± 7%  -7.78%  (p=0.002 n=10+9)
JumpdestOpAnalysis/PUSH2-6   1.32ms ± 4%   1.26ms ± 4%  -5.00%  (p=0.000 n=9+9)

name                        old speed     new speed     delta
JumpdestOpAnalysis/PUSH1-6  783MB/s ±16%  845MB/s ± 6%  +7.99%  (p=0.002 n=10+9)
JumpdestOpAnalysis/PUSH2-6  930MB/s ± 4%  979MB/s ± 5%  +5.31%  (p=0.000 n=9+9)
```